requirements.txt
Flask==3.0.0
vaderSentiment==3.3.2
google-generativeai==0.4.1
python-dotenv==1.0.1